added matrix

+++ utils.cpp
@@ -4,8 +4,8 @@
 
 #include "utils.h"
 
-uint64_t factorial(uint32_t num) {
-    uint64_t tmp=1;
+double factorial(double num) {
+    double tmp=1;
     for(auto i=2; i<=num ; i++){
         tmp*=i;
     }
