+++ utils.h
@@ -5,5 +5,5 @@
 #ifndef RFID_UTILS_H
 #define RFID_UTILS_H
 #include <iostream>
-uint64_t factorial(uint32_t num);
+double factorial(double num);
 #endif //RFID_UTILS_H
